avoid situation of that r is too big
previously if r is too big, it will not satisfy d2.size>r, and in turn the dds subroutine will not be inovked. while if r is too small, we may not get r combination. In another words, it may get traped to find useful r.  So the change is to avoid the situation of that r is too big. if d2.size<=r, we use d2.size-1 as the r such that the dds subroutine will be invoked.